requirements.txt
pydantic>=2.0
requests>=2.28
langchain-core>=0.3
langgraph>=0.2
langchain-openai>=0.2
python-dotenv>=1.0
